+++ src/Seccenter20240508/Model/GetRegularCleanResponse.php
@@ -38,7 +38,7 @@ class GetRegularCleanResponse implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'clean_days' => 'int32',
+        'clean_days' => 'int64',
         'switch' => null
     ];
