+++ src/Seccenter20240508/Model/KillProcListForGetHidsAlarmSummaryInfoOutput.php
@@ -39,7 +39,7 @@ class KillProcListForGetHidsAlarmSummaryInfoOutput implements ModelInterface, Ar
       */
     protected static $swaggerFormats = [
         'cmdline' => null,
-        'pid' => 'int32'
+        'pid' => 'int64'
     ];
 
     /**
