+++ src/Seccenter20240508/Model/ListAutoDefenseRulesRequest.php
@@ -41,8 +41,8 @@ class ListAutoDefenseRulesRequest implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'page_number' => 'int32',
-        'page_size' => 'int32',
+        'page_number' => 'int64',
+        'page_size' => 'int64',
         'sort_by' => null,
         'sort_order' => null,
         'top_group_id' => null
