+++ src/Seccenter20240508/Model/ListWhiteListsRequest.php
@@ -43,8 +43,8 @@ class ListWhiteListsRequest implements ModelInterface, ArrayAccess
       */
     protected static $swaggerFormats = [
         'conditions' => null,
-        'page_number' => 'int32',
-        'page_size' => 'int32',
+        'page_number' => 'int64',
+        'page_size' => 'int64',
         'sort_by' => null,
         'sort_order' => null,
         'type' => null
