+++ src/Seccenter20240508/Model/ResultForUpdateCommonAlarmStatusOutput.php
@@ -40,7 +40,7 @@ class ResultForUpdateCommonAlarmStatusOutput implements ModelInterface, ArrayAcc
       */
     protected static $swaggerFormats = [
         'alarm_id' => null,
-        'code' => 'int32',
+        'code' => 'int64',
         'msg' => null
     ];
